:shirt: Improving Gradio interface (shows pdf most cited by chunks)

+++ enstrag/data/parser.py
@@ -43,10 +43,9 @@ class Parser:
                 texts.append(str(elem))
 
         return " ".join(texts)
-        return [Document(page_content=text, metadata={"hash": sha256(text.encode('utf-8')).hexdigest(), "name": filename})]
 
     @staticmethod
-    def get_text_from_pdf(path_to_pdf: str, name: str = None) -> str:
+    def get_text_from_pdf(path_to_pdf: str) -> str:
         if not os.path.exists(path_to_pdf):
             raise FileNotFoundError(f"File {path_to_pdf} does not exist")
         if not path_to_pdf.endswith(".pdf"):
@@ -62,33 +61,40 @@ class Parser:
 
         return " ".join(texts)
 
-        return [Document(page_content=text, metadata={"hash": sha256(text.encode('utf-8')).hexdigest(), "name": name})]
-
     @staticmethod
     def get_text_from_pdf_url(url: str, name: str = None) -> str:
-        TMP_FOLDER = "/tmp/enstrag_"+str(pwd.getpwuid(os.getuid())[0])
+        if os.environ.get("PERSIST_PATH") is not None:
+            TMP_FOLDER = os.path.join(os.environ.get("PERSIST_PATH"), "pdfs") 
+        else:
+            TMP_FOLDER = "/tmp/enstrag_"+str(pwd.getpwuid(os.getuid())[0])
+
         os.makedirs(TMP_FOLDER, exist_ok=True)
         if name is None:
-            name = url
+            name = url.replace("/", "_")
+        name = name.replace(" ", "_")
 
-        with open(os.path.join(TMP_FOLDER, 'tmp.pdf'), 'wb') as f:
-            try:
-                response = requests.get(url)
-                f.write(response.content)
-            except Exception:
-                print(f"Failed to download {url}. Ignoring...")
-                return ""
+        pdf_path = os.path.join(TMP_FOLDER, f'{name}.pdf')
+        if not os.path.exists(pdf_path):
+            with open(pdf_path, 'wb') as f:
+                try:
+                    response = requests.get(url)
+                    f.write(response.content)
+                except Exception:
+                    print(f"Failed to download {url}. Ignoring...")
+                    return ""
 
-        text = Parser.get_text_from_pdf(os.path.join(TMP_FOLDER, 'tmp.pdf'), name=name)
-        shutil.rmtree(TMP_FOLDER)
-        return text
+        text = Parser.get_text_from_pdf(pdf_path)
+
+        if os.environ.get("PERSIST_PATH") is None:
+            shutil.rmtree(TMP_FOLDER)
+        return text, pdf_path
 
     @staticmethod
     def get_document_from_filedoc(filedoc: FileDocument) -> Document:
-        text =  Parser.get_text_from_pdf_url(filedoc.url, filedoc.name)
+        text, pdf_path =  Parser.get_text_from_pdf_url(filedoc.url, filedoc.name)
         if text == "":
             return None
-        return Document(page_content=text, metadata={"hash": sha256(text.encode('utf-8')).hexdigest(), "name": filedoc.name, "label": filedoc.label, "url": filedoc.url})
+        return Document(page_content=text, metadata={"hash": sha256(text.encode('utf-8')).hexdigest(), "name": filedoc.name, "label": filedoc.label, "url": filedoc.url, "path": pdf_path})
     
     @staticmethod
     def get_documents_from_filedocs(filedocs: List[FileDocument]) -> List[Document]:
